pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brady-gun-analysis"
version = "0.1.0"
description = "Crime Gun Supply Chain Analysis Dashboard for Brady Gun Center"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Brady Gun Center"}
]
keywords = ["gun-violence", "data-analysis", "etl", "dashboard", "streamlit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Legal Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openpyxl>=3.1.0",
    "streamlit>=1.28.0",
    "plotly>=5.18.0",
    "termcolor>=2.3.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
google = [
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
    "gspread>=5.12.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pandas-stubs>=2.0.0",
]
all = [
    "brady-gun-analysis[google,dev]",
]

[project.scripts]
brady-etl = "brady.etl.process_gunstat:main"
brady-dashboard = "brady.dashboard.app:main"

[project.urls]
Documentation = "docs/"
Repository = "https://github.com/bradygunproject/brady-gun-analysis"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
brady = ["py.typed"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
